// src/auth/admin/admin.dto.ts
import { IsEmail, IsString, MinLength } from 'class-validator';

export class AdminLoginDto {
  @IsEmail()
  email: string;

  @IsString()
  @MinLength(6)
  password: string;
}

export class CreateAdminDto {
  @IsEmail()
  email: string;

  @IsString()
  @MinLength(6)
  password: string;

  @IsString()
  name: string;
}

export class AdminResponseDto {
  id: string;
  email: string;
  name: string;
  role: string;
  isActive: boolean;
  accessToken: string;
}